# Handle cached byte array directly rather than reparsing