As per title. If cached, the bytes are readily available to be mucked with. No need to reparse necessarily. This spares a full copy for instance.

## Benchmark

```
name                                                                                old time/op    new time/op    delta
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-16    16.7µs ± 1%    16.5µs ± 1%  -1.57%  (p=0.008 n=5+5)
Unmarshal/base64_json_encoded_data_v1.0-16                                            11.2µs ± 1%    11.3µs ± 0%  +0.51%  (p=0.032 n=5+5)
Unmarshal/nil_data_v0.3-16                                                            14.7µs ± 1%    14.7µs ± 1%    ~     (p=0.841 n=5+5)
Unmarshal/struct_data_v1.0-16                                                         16.1µs ± 1%    16.1µs ± 1%    ~     (p=0.548 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-16    16.6µs ± 0%    16.3µs ± 1%  -2.09%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-16                                                         15.7µs ± 1%    15.8µs ± 1%    ~     (p=0.056 n=5+5)
Unmarshal/base64_xml_encoded_data_v1.0-16                                             11.7µs ± 1%    11.8µs ± 2%    ~     (p=0.746 n=5+5)
Unmarshal/nil_data_v1.0-16                                                            15.0µs ± 0%    15.0µs ± 1%    ~     (p=0.690 n=5+5)
Unmarshal/struct_data_v0.3-16                                                         16.4µs ± 1%    15.9µs ± 1%  -3.10%  (p=0.008 n=5+5)
Unmarshal/string_data_v0.3-16                                                         15.8µs ± 1%    15.5µs ± 1%  -2.13%  (p=0.008 n=5+5)

name                                                                                old alloc/op   new alloc/op   delta
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-16    3.40kB ± 0%    3.22kB ± 0%  -5.18%  (p=0.008 n=5+5)
Unmarshal/base64_json_encoded_data_v1.0-16                                            2.66kB ± 0%    2.66kB ± 0%    ~     (all equal)
Unmarshal/nil_data_v0.3-16                                                            3.14kB ± 0%    3.14kB ± 0%    ~     (all equal)
Unmarshal/struct_data_v1.0-16                                                         3.20kB ± 0%    3.20kB ± 0%    ~     (all equal)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-16    3.40kB ± 0%    3.22kB ± 0%  -5.18%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-16                                                         3.18kB ± 0%    3.18kB ± 0%    ~     (all equal)
Unmarshal/base64_xml_encoded_data_v1.0-16                                             2.67kB ± 0%    2.67kB ± 0%    ~     (all equal)
Unmarshal/nil_data_v1.0-16                                                            3.14kB ± 0%    3.14kB ± 0%    ~     (all equal)
Unmarshal/struct_data_v0.3-16                                                         3.37kB ± 0%    3.20kB ± 0%  -5.22%  (p=0.008 n=5+5)
Unmarshal/string_data_v0.3-16                                                         3.36kB ± 0%    3.18kB ± 0%  -5.24%  (p=0.008 n=5+5)

name                                                                                old allocs/op  new allocs/op  delta
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-16      69.0 ± 0%      67.0 ± 0%  -2.90%  (p=0.008 n=5+5)
Unmarshal/base64_json_encoded_data_v1.0-16                                              48.0 ± 0%      48.0 ± 0%    ~     (all equal)
Unmarshal/nil_data_v0.3-16                                                              63.0 ± 0%      63.0 ± 0%    ~     (all equal)
Unmarshal/struct_data_v1.0-16                                                           66.0 ± 0%      66.0 ± 0%    ~     (all equal)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-16      69.0 ± 0%      67.0 ± 0%  -2.90%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-16                                                           65.0 ± 0%      65.0 ± 0%    ~     (all equal)
Unmarshal/base64_xml_encoded_data_v1.0-16                                               48.0 ± 0%      48.0 ± 0%    ~     (all equal)
Unmarshal/nil_data_v1.0-16                                                              63.0 ± 0%      63.0 ± 0%    ~     (all equal)
Unmarshal/struct_data_v0.3-16                                                           68.0 ± 0%      66.0 ± 0%  -2.94%  (p=0.008 n=5+5)
Unmarshal/string_data_v0.3-16                                                           67.0 ± 0%      65.0 ± 0%  -2.99%  (p=0.008 n=5+5)
```